Create SQL - Query to summarize users activities
Here's the SQL query used to retrieve the sports data of the platform's users over the last 90 days. The aim is to set up a content recommendation system. To do this, we're looking for key information, in weekly, monthly or global averages, for the various main sports (walking, running, cycling, swimming, workout). For speeds and durations, on a global scale, I preferred to calculate medians, as the average values were regularly altered by outliers due to the database.