Clustering to define users interest
In this approach, we use the K-Means method, to cluster users based on their similarities. Clustering using K-Means is a method of dividing a set of data points into groups or clusters based on their similarities. The algorithm works by identifying k number of centroids, or cluster centers, and assigning each data point to the nearest centroid. The centroids are then updated based on the mean value of all the points assigned to that cluster, and the process is repeated until the clusters no longer change significantly.

The K-Means algorithm is a powerful tool that helps us group data points based on their similarities. This approach allows us to identify different groups or patterns in the data, which can be used to push personalized content to users based on their level in each of the sports.

To perform this clustering, we use several data: for running, cycling, swimming, and walking, the features are weekly distance and median speed. For yoga and workout, these are the average duration and the number of average monthly sessions. The SQL request returns a line for each user who has been active in the last 3 months with each of these features and the value of the user, even if it is zero. From this matrix, we perform the clustering. This brings us back for each user and for each sport a value between 0 and 3. The choice of a range from 0 to 3 is not random. To achieve this result, we carried out the elbow method for each activity, which brings us back the appropriate number of clusters for each sport, which gave us this distribution for each.

Before carrying out the clustering, many preliminary steps were necessary. A large part of the work in AI is cleaning and preparing data because the raw data provided by the SQL request did not allow for an effective clustering algorithm.

Initially, we had to abandon the idea of average speed and replace it with median speed. This measure reduces the impact of erroneous speed measures by tracking applications, which can distort the analysis.

Then, for each clustering (1 per sport), we excluded users with zeros in any of the columns. This was because the large number of zero values had a considerable effect on the analysis. These users were excluded from the cluster and assigned a 'null' value as a cluster.