Add files via upload
In this presentation, you'll find some of the results of conversation analysis, carried out using the python script of the same name. The purpose of this presentation was to present technical analyses to my product team, which meant I had to focus on clear results that everyone could understand. It was also important for me to raise the business stakes of this study, by proposing concrete actions to be taken in order to impact the figures presented. In this way, during this project, I was able to put my technical study at the service of the company's business issues, so as not to lose sight of the importance of carrying out these studies, an issue that is particularly close to my heart.