Add files via upload
You will find here the long presentation of the running 2022 observatory. All numerical results come from the SQL analyses I performed using the SQL script of the same name. These analyses enabled me to carry out a complete and relevant study of runners' habits in 2022, which was then presented to all the major players in running in France, whether commercial or journalistic. It's a study I'm particularly proud of, as it has had a real impact on the decision-making of many players in the sport.