Add files via upload
You will find here the long presentation of the Observatoire du bien être au travail 2023. All the quantified results come from numerous Excel analyses I carried out. These analyses have enabled me to carry out a complete and relevant study of the current situation of workers in France, from a well-being point of view. The aim was to gain a better understanding of employees' expectations, the levers for mobilizing them and the actions to be taken to retain them within the company. The study was presented to over 100 corporate decision-makers, human resources professionals and journalists.